Refactored app creation and updating
